Added Linux Support and some formatting
Added final keywords to a lot of local variables and other minor code changes. Also removed Windows-style line endings from the file. The background service functionality still needs to be implemented for Linux, probably using cron jobs in a manner that is analogous to what's being done with the method that creates the batch file right now.